refactor(operators): remove circular reference using ambient merge
- remove circular references using
  https://github.com/Microsoft/TypeScript/issues/3332, parents are
references children via ambient interfaces and let compiler merges it
