+++ src/ambient/ConnectableObservable.ts
@@ -1,0 +1,1 @@
+export interface ConnectableObservable<T> {}
