+++ src/ambient/GroupedObservable.ts
@@ -1,0 +1,1 @@
+export interface GroupedObservable<T> {}
