+++ src/ambient/Subject.ts
@@ -1,0 +1,1 @@
+export interface Subject<T> {}
